Refactor : Rename src folder in order to have a more readable imports
