+++ src/resources/constants.js
